+++ src/config.ts
@@ -37,6 +37,7 @@ export interface Config {
   npmrc?: string;
   csvFileStats: string;
   csvFileRaw: string;
+  partition: 'none' | 'measurement';
 }
 
 export async function makeConfig(opts: Opts): Promise<Config> {
@@ -86,6 +87,9 @@ export async function makeConfig(opts: Opts): Promise<Config> {
     }
     if (opts['window-size'] !== undefined) {
       throw new Error('--window-size cannot be specified when using --config');
+    }
+    if (opts['partition'] !== undefined) {
+      throw new Error('--partition cannot be specified when using --config');
     }
     const rawConfigObj = await fsExtra.readJson(opts.config);
     const validatedConfigObj = await parseConfigFile(rawConfigObj, opts.config);
@@ -175,6 +179,8 @@ export function applyDefaults(partial: Partial<Config>): Config {
         : defaults.resolveBareModules,
     root: partial.root !== undefined ? partial.root : defaults.root,
     timeout: partial.timeout !== undefined ? partial.timeout : defaults.timeout,
+    partition:
+      partial.partition !== undefined ? partial.partition : defaults.partition,
   };
 }